git-svn-id: https://gentlenav.googlecode.com/svn/trunk@2145 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libUDB/magnetometerOptions.h
@@ -268,6 +268,6 @@
 #define MAGNETICMAXIMUM 1500
 
 // The following line computes an internal parameter, do not change it.
-#define DECLINATIONANGLE ((int)(MAGNETICDECLINATION *(32767.0 / 180.0)))
+#define DECLINATIONANGLE ((int16_t)(MAGNETICDECLINATION *(32767.0 / 180.0)))
