+++ libUDB/servoOut.c
@@ -75,7 +75,7 @@
 #define SCALE_FOR_PWM_OUT(x)		((x) << 1)
 #elif ( CLOCK_CONFIG == FRC8X_CLOCK )
 #define PWMOUTSCALE					60398	// = 256*256*(3.6864/4)
-#define SCALE_FOR_PWM_OUT(x)		(((union longww)(long)__builtin_muluu( (x) ,  PWMOUTSCALE ))._.W1)
+#define SCALE_FOR_PWM_OUT(x)		(((union longww)(int32_t)__builtin_muluu( (x) ,  PWMOUTSCALE ))._.W1)
 #endif
 
 #endif
@@ -83,14 +83,14 @@
 
 //	routines to drive the PWM pins for the servos,
 
-int udb_pwOut[NUM_OUTPUTS+1] ;	// pulse widths for servo outputs
+int16_t udb_pwOut[NUM_OUTPUTS+1] ;	// pulse widths for servo outputs
 
-int outputNum ;
+int16_t outputNum ;
 
 
 void udb_init_pwm( void )	// initialize the PWM
 {
-	int i;
+	int16_t i;
 	for (i=0; i <= NUM_OUTPUTS; i++)
 		udb_pwOut[i] = 0;
 	
@@ -164,7 +164,7 @@ void start_pwm_outputs( void )
 
 
 #if (RECORD_FREE_STACK_SPACE == 1)
-extern unsigned int maxstack ;
+extern uint16_t maxstack ;
 #endif
 
 
@@ -257,7 +257,7 @@ void __attribute__((__interrupt__,__no_auto_psv__)) _T4Interrupt(void)
 	// Check stack space here because it's a high-priority ISR
 	// which may have interrupted a whole chain of other ISRs,
 	// So available stack space can get lowest here.
-	unsigned int stack = WREG15 ;
+	uint16_t stack = WREG15 ;
 	if ( stack > maxstack )
 	{
 		maxstack = stack ;
